New feature: autotype to active window
This patch adds a new feature:
dmoji can now inject the emoji directly into the currently active
window.
This way the user does not need to manually paste the content of the
keyboard. It also helps to keep the clipboard buffer clean.

The associated option is `-t` (for Type it)